Move static file handling into own class
Moves multiple static file handling methods, such as FindFileX out of
ExperimentInfo into their own class
