+++ Framework/API/inc/MantidAPI/ExperimentInfo.h
@@ -147,23 +147,6 @@ public:
   // run end time if available, empty otherwise
   std::string getAvailableWorkspaceEndDate() const;
 
-  /// Utility to retrieve the validity dates for the given IDF
-  static void getValidFromTo(const std::string &IDFfilename,
-                             std::string &outValidFrom,
-                             std::string &outValidTo);
-  /// Utility to retrieve a resource file (IDF, Parameters, ..)
-  static std::vector<std::string> getResourceFilenames(
-      const std::string &prefix, const std::vector<std::string> &fileFormats,
-      const std::vector<std::string> &directoryNames, const std::string &date);
-  /// Get the IDF using the instrument name and date
-  static std::string getInstrumentFilename(const std::string &instrumentName,
-                                           const std::string &date = "");
-
-  /// Search instrument directories for Parameter file,
-  /// return full path name if found, else "".
-  static std::string getFullPathParamIDF(std::string instName,
-                                         const std::string &dirHint = "");
-
   const Geometry::DetectorInfo &detectorInfo() const;
   Geometry::DetectorInfo &mutableDetectorInfo();
