+++ Framework/Algorithms/src/VesuvioL1ThetaResolution.cpp
@@ -7,6 +7,7 @@
 #include "MantidAlgorithms/VesuvioL1ThetaResolution.h"
 
 #include "MantidAPI/AlgorithmManager.h"
+#include "MantidAPI/FileFinderUtils.h"
 #include "MantidAPI/FileProperty.h"
 #include "MantidAPI/MatrixWorkspace.h"
 #include "MantidAPI/SpectrumInfo.h"
@@ -267,7 +268,8 @@ void VesuvioL1ThetaResolution::loadInstrument() {
   // Get the filename for the VESUVIO IDF
   MatrixWorkspace_sptr tempWS =
       WorkspaceFactory::Instance().create("Workspace2D", 1, 1, 1);
-  const std::string vesuvioIPF = tempWS->getInstrumentFilename("VESUVIO");
+  const std::string vesuvioIPF =
+      FileFinderUtils::getInstrumentFilename("VESUVIO");
 
   // Load an empty VESUVIO instrument workspace
   IAlgorithm_sptr loadInst =
