+++ Framework/PythonInterface/mantid/api/src/Exports/ExperimentInfo.cpp
@@ -5,6 +5,7 @@
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
 #include "MantidAPI/ExperimentInfo.h"
+#include "MantidAPI/FileFinderUtils.h"
 #include "MantidAPI/Run.h"
 #include "MantidAPI/Sample.h"
 #include "MantidAPI/SpectrumInfo.h"
@@ -23,6 +24,7 @@
 #include <boost/python/register_ptr_to_python.hpp>
 
 using Mantid::API::ExperimentInfo;
+using Mantid::API::FileFinderUtils;
 using Mantid::PythonInterface::Policies::RemoveConstSharedPtr;
 using namespace Mantid::PythonInterface;
 using namespace boost::python;
@@ -33,7 +35,7 @@ GET_POINTER_SPECIALIZATION(ExperimentInfo)
 list getResourceFilenames(const std::string &prefix, const list &fileFormats,
                           const list &directoryNames, const std::string &date) {
   return Converters::ToPyList<std::string>()(
-      ExperimentInfo::getResourceFilenames(
+      FileFinderUtils::getResourceFilenames(
           prefix, Converters::PySequenceToVector<std::string>(fileFormats)(),
           Converters::PySequenceToVector<std::string>(directoryNames)(), date));
 }
@@ -44,7 +46,7 @@ GNU_DIAG_OFF("unused-local-typedef")
 GNU_DIAG_OFF("conversion")
 /// Overload generator for getInstrumentFilename
 BOOST_PYTHON_FUNCTION_OVERLOADS(getInstrumentFilename_Overload,
-                                ExperimentInfo::getInstrumentFilename, 1, 2)
+                                FileFinderUtils::getInstrumentFilename, 1, 2)
 GNU_DIAG_ON("conversion")
 GNU_DIAG_ON("unused-local-typedef")
 
@@ -64,7 +66,8 @@ void export_ExperimentInfo() {
       .def("getInstrument", &ExperimentInfo::getInstrument,
            return_value_policy<RemoveConstSharedPtr>(), args("self"),
            "Returns the :class:`~mantid.geometry.Instrument` for this run.")
-      .def("getResourceFilenames", &getResourceFilenames,
+      // FileFinderUtils binds Kept for backwards compat.
+      .def("getResourceFilenames", &FileFinderUtils::getResourceFilenames,
            (arg("prefix"), arg("fileFormats"), arg("directoryNames"),
             arg("date")),
            "Compile a list of files in compliance with name pattern-matching,\n"
@@ -80,7 +83,7 @@ void export_ExperimentInfo() {
            "file will encapsulate this date (e.g '1900-01-31 23:59:00')\n"
            "\nreturns : list of absolute paths for each valid file\n")
       .staticmethod("getResourceFilenames")
-      .def("getInstrumentFilename", &ExperimentInfo::getInstrumentFilename,
+      .def("getInstrumentFilename", &FileFinderUtils::getInstrumentFilename,
            getInstrumentFilename_Overload(
                "Returns IDF filename", (arg("instrument"), arg("date") = "")))
       .staticmethod("getInstrumentFilename")
