+++ qt/widgets/common/src/SlitCalculator.cpp
@@ -7,6 +7,7 @@
 #include "MantidQtWidgets/Common/SlitCalculator.h"
 #include "MantidAPI/AlgorithmManager.h"
 #include "MantidAPI/ExperimentInfo.h"
+#include "MantidAPI/FileFinderUtils.h"
 #include "MantidAPI/Progress.h"
 #include "MantidGeometry/Instrument/InstrumentDefinitionParser.h"
 #include "MantidKernel/Strings.h"
@@ -39,7 +40,7 @@ void SlitCalculator::setInstrument(const std::string &instrumentName) {
       Mantid::Types::Core::DateAndTime::getCurrentTime().toISO8601String();
   // find the full path to the definition file
   auto filename =
-      Mantid::API::ExperimentInfo::getInstrumentFilename(instrumentName, date);
+      Mantid::API::FileFinderUtils::getInstrumentFilename(instrumentName, date);
   // parse the XML that we have found for the definition
   Mantid::Geometry::InstrumentDefinitionParser parser =
       Mantid::Geometry::InstrumentDefinitionParser(
